Successfully implemented the following functionalities in both frontend and backend: - User registration - DONE - User login - DONE - Email verification for account activation - DONE - Email verification for password recovery - DONE - Authentication management with Passport.js and JWT - DONE - Password recovery - DONE

+++ public/js/controllers/auth.controller.js
@@ -1,12 +1,34 @@
-import { signUp } from '../api/auth.js';
+import { changePassword, signIn, signUp } from '../api/auth.js';
 import { resendVerificationEmail } from '../api/auth.js';
-import { renderRoute } from "../router.js";
+import { navigateTo, renderRoute } from "../router.js";
 
-export function handleSignUp() {
-  const form = document.getElementById('sign-up-form');
+export async function signInHandler(){
+  const userData = {
+    email: document.getElementById('email-sign-in').value,
+    password: document.getElementById('password-sign-in').value,
+    rememberMe: document.getElementById('remember-session').checked,
+  };
+  console.log(userData.rememberMe);
+  try {
+    const validation = validatePassword(userData.password);
+    if (!validation.valid) {
+      alert(validation.message);
+    }
+    const { accessToken } = await signIn(userData);
+    if(accessToken){
+      window.accessToken = accessToken;
+      alert('¡Email verified successfully!');
+      navigateTo('/project-screen');
+    }else {
+      alert('Error logging');
+    }
+  } catch (error) {
+    console.error('Sign up error:', error);
+    alert(error.message || 'Error logging');
+  }
+}
 
-  form.addEventListener('submit', async (event) => {
-    event.preventDefault();
+export async function handleSignUp() {
     const userData = {
       name: document.getElementById('name-sign-up').value,
       email: document.getElementById('email-sign-up').value,
@@ -14,21 +36,23 @@ export function handleSignUp() {
       confirmPassword: document.getElementById('confirm-password-sign-up').value,
     };
     try {
+      if (userData.password !== userData.confirmPassword) {
+        alert('The passwords do not match');
+        return;
+      }
       const validation = validatePassword(userData.password);
       if (!validation.valid) {
         alert(validation.message);
+        return;
       }
       const result = await signUp(userData);
       console.log('Backend response:', result);
-
-      window.history.pushState(null, '', '/sign-up/verify-email');
-      renderRoute('/sign-up/verify-email');
+      navigateTo('/sign-up/verify-email');
       localStorage.setItem('email', userData.email);
     } catch (error) {
       console.error('Sign up error:', error);
-      alert(error.message);
+      alert(`Error: ${error.message}`);
     }
-  });
 }
 
 function validatePassword(password){
@@ -42,9 +66,7 @@ function validatePassword(password){
   return { valid: true, message: "Valide password." };
 }
 
-
-export async function handleResendVerificationEmail(){
-  const email = localStorage.getItem('email');
+export async function handleResendVerificationEmail(email){
   if (!email) {
     alert('Email not found. Please try again.');
     return;
@@ -52,7 +74,61 @@ export async function handleResendVerificationEmail(){
   try {
     const result = await resendVerificationEmail(email);
     alert(result.message || 'Verification email resent successfully!');
+    return result;
   } catch (error) {
-    alert('Failed to resend verification email. Please try again later.');
+    if (error.message.includes('Too many email requests')) {
+      alert('You have exceeded the maximum number of attempts. Please try again later.');
+    } else {
+      alert('Failed to resend verification email. Please try again later.');
+    }
   }
+}
+
+export async function sendEmailToChangePassword(){
+  const form = document.getElementById('password-recovery-form');
+  const userEmail = document.getElementById('email-sign-in').value;
+  if (!userEmail) {
+    alert('You need to write an email.');
+  }
+  try {
+    const response = await handleResendVerificationEmail(userEmail);
+    if(response){
+      localStorage.setItem('email', userEmail);
+      localStorage.setItem('state', 'change-password');
+      navigateTo('/sign-in/recovery-password/verify-email');
+    }else {
+      alert('Error send email. Try again');
+    }
+  } catch (error) {
+    console.error('Internal server error');
+    alert(error.message || 'Error logging');
+  }
+}
+
+export async function changePasswordHandler(){
+  const form = document.getElementById('change-password-form');
+
+  form.addEventListener('submit', async (event) => {
+    event.preventDefault();
+    const userData = {
+      newPassword: document.getElementById('new-password').value,
+      confirmNewPassword: document.getElementById('confirm-new-password').value,
+    };
+    try {
+      const validation = validatePassword(userData.newPassword);
+      if (!validation.valid) {
+        alert(validation.message);
+      }
+      const response = await changePassword(userData);
+      if(response){
+        alert('¡Password changed successfully! Please, sign in.');
+        navigateTo('/sign-in');
+      }else {
+        alert('Error changing password');
+      }
+    } catch (error) {
+      console.error('Error changing password:', error);
+      alert(error.message || 'Error changing password');
+    }
+  });
 }
